boot_reqs/requirements.txt
yara
pwntools
pylzma
